requirements.txt
langchain
langchain-ollama
ollama

pyaudio
wave

openai-whisper


speechrecognition

edge-tts

playsound3
